moved vehicles to lib
