+++ src/components/vehicleType/vehicleTypeModal.tsx
@@ -1,9 +1,9 @@
 import { useVehicleTypeContext } from "@/util/context/VehicleTypeContext";
-import { defaultCopter, defaultPlane } from "@/util/defaultVehicles";
 import Button from "../toolBar/button";
 import DraggableNumberInput from "../ui/draggableNumericInput";
 import { getMinTurnRadius } from "@/lib/dubins/dubinWaypoints";
 import { cn } from "@/lib/utils";
+import { defaultCopter, defaultPlane } from "@/lib/vehicles/defaults";
 
 export default function VehicleTypeModal() {
   let { vehicle, setVehicle } = useVehicleTypeContext()
